refactor: change the way blog card hover is implemented, optimize it like tech-list
